requirements/ci.txt
# SHA1:6eb3d9968c46af6d37be7fa6f15da31abb84b76c
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
astroid==2.12.13
    # via
    #   -c requirements/local.txt
    #   pylint
attrs==22.1.0
    # via
    #   -c requirements/local.txt
    #   pytest
blinker==1.8.2
    # via
    #   -c requirements/local.txt
    #   flask
build==0.9.0
    # via
    #   -c requirements/local.txt
    #   pytest-checkdocs
click==8.1.3
    # via
    #   -c requirements/local.txt
    #   flask
coverage[toml]==6.5.0
    # via
    #   -c requirements/local.txt
    #   pytest-cov
docutils==0.19
    # via
    #   -c requirements/local.txt
    #   pytest-checkdocs
exceptiongroup==1.0.4
    # via
    #   -c requirements/local.txt
    #   pytest
flake8==6.0.0
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
    #   flake8-pyproject
flake8-pyproject==1.2.2
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
flask==3.0.3
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
importlib-metadata==5.1.0
    # via
    #   -c requirements/local.txt
    #   flask
    #   pytest-checkdocs
iniconfig==1.1.1
    # via
    #   -c requirements/local.txt
    #   pytest
isort==5.10.1
    # via
    #   -c requirements/local.txt
    #   pylint
itsdangerous==2.2.0
    # via
    #   -c requirements/local.txt
    #   flask
jinja2==3.1.2
    # via
    #   -c requirements/local.txt
    #   flask
lazy-object-proxy==1.8.0
    # via
    #   -c requirements/local.txt
    #   astroid
markupsafe==2.1.1
    # via
    #   -c requirements/local.txt
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via
    #   -c requirements/local.txt
    #   flake8
    #   pylint
mypy==0.991
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
mypy-extensions==0.4.3
    # via
    #   -c requirements/local.txt
    #   mypy
nodeenv==1.7.0
    # via
    #   -c requirements/local.txt
    #   pyright
packaging==21.3
    # via
    #   -c requirements/local.txt
    #   build
    #   pytest
pep517==0.13.0
    # via
    #   -c requirements/local.txt
    #   build
platformdirs==2.5.4
    # via
    #   -c requirements/local.txt
    #   pylint
pluggy==1.0.0
    # via
    #   -c requirements/local.txt
    #   pytest
pycodestyle==2.10.0
    # via
    #   -c requirements/local.txt
    #   flake8
pyflakes==3.0.1
    # via
    #   -c requirements/local.txt
    #   flake8
pylint==2.15.8
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
pyparsing==3.0.9
    # via
    #   -c requirements/local.txt
    #   packaging
pyright==1.1.380
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
pytest==7.2.0
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
    #   pytest-cov
pytest-checkdocs==2.9.0
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
pytest-cov==4.0.0
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
requests-mock==1.12.1
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
tomli==2.0.1
    # via
    #   -c requirements/local.txt
    #   build
    #   coverage
    #   flake8-pyproject
    #   mypy
    #   pylint
    #   pytest
tomlkit==0.11.6
    # via
    #   -c requirements/local.txt
    #   pylint
types-requests==2.28.11.5
    # via
    #   -c requirements/local.txt
    #   -r requirements/ci.in
types-urllib3==1.26.25.4
    # via
    #   -c requirements/local.txt
    #   types-requests
typing-extensions==4.4.0
    # via
    #   -c requirements/local.txt
    #   astroid
    #   mypy
    #   pylint
werkzeug==3.0.4
    # via
    #   -c requirements/local.txt
    #   flask
wrapt==1.14.1
    # via
    #   -c requirements/local.txt
    #   astroid
zipp==3.11.0
    # via
    #   -c requirements/local.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==65.6.3
    # via
    #   -c requirements/local.txt
    #   nodeenv
